chlopaki nie farmie commitow tylko chce na kompie spullowac brancha
